NewsRoomAI-Co-Writer-main/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
passlib[bcrypt]
python-jose
python-dotenv
pydantic[email]

